pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis-voice-assistant"
version = "2.0.0"
description = "A professional voice assistant powered by local LLMs with proper software engineering architecture"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Jarvis Team", email = "jarvis@example.com"},
]
keywords = [
    "voice-assistant",
    "ai",
    "llm",
    "speech-recognition",
    "text-to-speech",
    "ollama",
    "langchain"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "langchain>=0.1.0",
    "langchain-ollama>=0.1.0",
    "speechrecognition>=3.10.0",
    "pyttsx3>=2.90",
    "pyaudio>=0.2.11",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

[project.urls]
Homepage = "https://github.com/Prophet60191/jarvis"
Repository = "https://github.com/Prophet60191/jarvis.git"

[project.scripts]
jarvis = "jarvis.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["jarvis*"]
exclude = ["tests*", "archive*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]
